+++ Classes/Indexer/RecordIndexer.php
@@ -1,10 +1,12 @@
 <?php
+
 namespace SourceBroker\Hugo\Indexer;
 
+use SourceBroker\Hugo\Configuration\Configurator;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
 use TYPO3\CMS\Core\Database\ConnectionPool;
-use SourceBroker\Hugo\Domain\Model\Document;
 use SourceBroker\Hugo\Domain\Model\DocumentCollection;
+use TYPO3\CMS\Extbase\Object\ObjectManager;
 
 class RecordIndexer extends AbstractIndexer
 {
@@ -13,36 +15,45 @@ class RecordIndexer extends AbstractIndexer
      * @param DocumentCollection $documentCollection
      * @return array
      */
-    public function runCollection(int $pageUid, DocumentCollection $documentCollection): array
+    public function getDocumentsForPage(int $pageUid, DocumentCollection $documentCollection): array
     {
-        $table = 'tx_news_domain_model_news';
+        $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
+        $hugoConfig = $objectManager->get(Configurator::class, null, $pageUid);
 
-        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable($table);
-        $queryBuilder->select('*')->from($table)->where(
-            $queryBuilder->expr()->eq('pid', $queryBuilder->createNamedParameter($pageUid, \PDO::PARAM_INT))
-        );
-        $recordRows = $queryBuilder->execute()->fetchAll();
+        foreach($hugoConfig->getOption('indexer.records.exporter') as $exporterConfig) {
+            if ($pageUid == $exporterConfig['pageUid']) {
+                $table = $exporterConfig['table'];
+                $recordsPid = $exporterConfig['recordsPid'];
+                $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable($table);
+                $queryBuilder->select('*')->from($table)->where(
+                    $queryBuilder->expr()->eq('pid', $queryBuilder->createNamedParameter($recordsPid, \PDO::PARAM_INT))
+                );
+                $recordRows = $queryBuilder->execute()->fetchAll();
 
-        foreach ($recordRows as $record) {
-            $document = $documentCollection->create();
+                foreach ($recordRows as $record) {
+                    $slug = $this->slugify($record['title']);
+                    $document = $documentCollection->create();
+                    $document->setStoreFilename($record['uid'] . '_' . ucfirst($slug))
+                        ->setId($record['uid'])
+                        ->setTitle($record['title'])
+                        ->setSlug($slug);
 
-            $document->setId($record['uid']);
-            $document->setTitle($record['title']);
-            $document->setSlug($this->slugify($record['title']));
-
-            //$document->setDescription($record['description']);
-            //$document->setKeywords($record['keywords']);
-            //$document->setDate($record['datetime']);
-            //$document->setTeaser($record['teaser']);
-            //$document->setBodytext($record['bodytext']);
-            //$document->setDraft($record['hidden']);
-            //$document->setEndtime($record['expirydate']);
+                    //$document->setDescription($record['description']);
+                    //$document->setKeywords($record['keywords']);
+                    //$document->setDate($record['datetime']);
+                    //$document->setTeaser($record['teaser']);
+                    //$document->setBodytext($record['bodytext']);
+                    //$document->setDraft($record['hidden']);
+                    //$document->setEndtime($record['expirydate']);
+                }
+            }
         }
 
         return [
             $pageUid,
             $documentCollection
         ];
+
     }
 
 }