+++ Classes/Traversing/TreeTraverser.php
@@ -2,11 +2,13 @@
 
 namespace SourceBroker\Hugo\Traversing;
 
-use SourceBroker\Hugo\Domain\Model\Document;
+use Cocur\Slugify\Slugify;
 use SourceBroker\Hugo\Domain\Model\DocumentCollection;
-use TYPO3\CMS\Core\Database\ConnectionPool;
+use SourceBroker\Hugo\Domain\Repository\Typo3PageRepository;
 use TYPO3\CMS\Core\Utility\GeneralUtility;
+use TYPO3\CMS\Extbase\Object\ObjectManager;
 use TYPO3\CMS\Extbase\SignalSlot\Dispatcher;
+use TYPO3\CMS\Frontend\Page\PageRepository;
 
 /**
  * Class PageTraverser
@@ -26,7 +28,6 @@ class TreeTraverser
         $this->writer->clean();
     }
 
-
     /**
      * @param int $pageUid
      * @param string[] $path
@@ -35,55 +36,29 @@ class TreeTraverser
      */
     public function start(int $pageUid, array $path = []): void
     {
-        $signalDispatcher = GeneralUtility::makeInstance(Dispatcher::class);
+        $objectManager = GeneralUtility::makeInstance(ObjectManager::class);
+        $signalDispatcher = $objectManager->get(Dispatcher::class);
+        $slugifier = $objectManager->get(Slugify::class);
+        $typo3PageRepository = $objectManager->get(Typo3PageRepository::class);
+        $typo3Page = $typo3PageRepository->getByUid($pageUid);
 
-        $document = new Document();
-        $document->setType(Document::TYPE_PAGE);
-
-        $signalDispatcher->dispatch(__CLASS__, 'document', [
-            $pageUid,
-            $document,
-        ]);
-
-        if ($document->getDeleted()) {
-            return;
+        if(!$typo3Page['is_siteroot'] && $typo3Page['doktype'] !== PageRepository::DOKTYPE_SYSFOLDER) {
+            $path[] = $slugifier->slugify($typo3Page['nav_title'] ?: $typo3Page['title']);
         }
 
-        if (!$document->isRoot()) {
-            $path[] = $document->getSlug();
-        }
-
-        $this->writer->save($document, $path);
-
-        $documentCollection = new DocumentCollection();
-
-        $signalDispatcher->dispatch(__CLASS__, 'extractDocuments', [
+        $documentCollection = $objectManager->get(DocumentCollection::class);
+        $signalDispatcher->dispatch(__CLASS__, 'getDocumentsForPage', [
             $pageUid,
             $documentCollection,
         ]);
-
         $this->writer->saveDocuments($documentCollection, $path);
 
-        $pages = $this->getPages($pageUid);
-
-        foreach ($pages as $page) {
+        foreach ($typo3PageRepository->getPagesByPidAndDoktype($pageUid, [
+            PageRepository::DOKTYPE_DEFAULT,
+            PageRepository::DOKTYPE_SYSFOLDER,
+            PageRepository::DOKTYPE_SHORTCUT,
+        ]) as $page) {
             $this->start($page['uid'], $path);
         }
-    }
-
-    /**
-     * @param int $pid
-     * @return array
-     */
-    protected function getPages(int $pid)
-    {
-        $queryBuilder = GeneralUtility::makeInstance(ConnectionPool::class)->getQueryBuilderForTable('pages');
-        $queryBuilder->select('uid', 'title')->from('pages')->where(
-            $queryBuilder->expr()->eq('pid', $queryBuilder->createNamedParameter($pid, \PDO::PARAM_INT)),
-            $queryBuilder->expr()->in('doktype', $queryBuilder->createNamedParameter('1,2', \PDO::PARAM_INT))
-
-        );
-
-        return $queryBuilder->execute()->fetchAll();
     }
 }